// frontend/src/api/models/SubtitlePosition.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SubtitlePosition = {
    Top: 'top',
    Bottom: 'bottom',
    Middle: 'middle',
    Left: 'left',
    Right: 'right',
    Center: 'center',
    TopLeft: 'top_left',
    TopRight: 'top_right',
    BottomLeft: 'bottom_left',
    BottomRight: 'bottom_right',
    TopLeftOfTwo: 'top_left_of_two',
    TopRightOfTwo: 'top_right_of_two',
    BottomLeftOfTwo: 'bottom_left_of_two',
    BottomRightOfTwo: 'bottom_right_of_two',
    TopLeftOfThree: 'top_left_of_three',
    TopRightOfThree: 'top_right_of_three',
    BottomLeftOfThree: 'bottom_left_of_three',
    BottomRightOfThree: 'bottom_right_of_three',
    TopLeftOfFour: 'top_left_of_four',
    TopRightOfFour: 'top_right_of_four',
    BottomLeftOfFour: 'bottom_left_of_four',
    BottomRightOfFour: 'bottom_right_of_four',
    TopLeftOfFive: 'top_left_of_five',
    TopRightOfFive: 'top_right_of_five',
    BottomLeftOfFive: 'bottom_left_of_five',
    BottomRightOfFive: 'bottom_right_of_five'
} as const;
export type SubtitlePosition = typeof SubtitlePosition[keyof typeof SubtitlePosition];


export function instanceOfSubtitlePosition(value: any): boolean {
    for (const key in SubtitlePosition) {
        if (Object.prototype.hasOwnProperty.call(SubtitlePosition, key)) {
            if (SubtitlePosition[key as keyof typeof SubtitlePosition] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SubtitlePositionFromJSON(json: any): SubtitlePosition {
    return SubtitlePositionFromJSONTyped(json, false);
}

export function SubtitlePositionFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubtitlePosition {
    return json as SubtitlePosition;
}

export function SubtitlePositionToJSON(value?: SubtitlePosition | null): any {
    return value as any;
}

export function SubtitlePositionToJSONTyped(value: any, ignoreDiscriminator: boolean): SubtitlePosition {
    return value as SubtitlePosition;
}